core: create tedi_core, move leaf primitives (locate/marker/error/blockers)
- new tedi_core crate: tedi_md -> tedi_core -> {tedi, tedi_adapters}
- RepoInfo moves adapters->core; IssueLink/IssueSelector/IssueIndex/IssueRef
  consolidate into core/locate.rs; marker.rs, error.rs move verbatim
- BlockerSequence renamed Blockers (BlockerSequenceExt -> BlockersExt); dead
  join_with_blockers deleted (Issue::body covers it)
- current_user lives in core; From<&IssueIdentity> for IssueMarker parks in
  tedi/types.rs until IssueIdentity follows in stage 3a
