+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "3"
-members = ["tedi", "tedi_adapters", "tedi_eval", "tedi_md"]
+members = ["tedi", "tedi_adapters", "tedi_core", "tedi_eval", "tedi_md"]
 
 [workspace.package]
 version = "0.16.3"
@@ -11,6 +11,7 @@ repository = "https://github.com/valeratrades/tedi"
 
 [workspace.dependencies]
 tedi_adapters = { path = "tedi_adapters", version = "0.16.3" }
+tedi_core = { path = "tedi_core", version = "0.16.3" }
 tedi_eval = { path = "tedi_eval", version = "0.16.3" }
 tedi_md = { path = "tedi_md", version = "0.16.3" }
