+++ tedi/src/blocker_interactions/operations.rs
@@ -1,14 +1,14 @@
 //! Extended stack operations for blocker management.
 //!
-//! This module provides extension methods on BlockerSequence from the library:
+//! This module provides extension methods on Blockers from the library:
 //! - `add`: Push a new blocker onto the stack
 //! - `pop`: Remove the last blocker from the stack
 //! - `current`: Get the current (last) blocker with its parent context
 
-use tedi::{BlockerItem, BlockerSequence, IssueRef};
+use tedi::{BlockerItem, Blockers, IssueRef};
 
-/// Extension trait for BlockerSequence with additional operations
-pub trait BlockerSequenceExt {
+/// Extension trait for Blockers with additional operations
+pub trait BlockersExt {
 	/// Get the current (last) blocker item
 	fn current(&self) -> Option<&BlockerItem>;
 
@@ -42,7 +42,7 @@ pub trait BlockerSequenceExt {
 	fn set(&mut self, text: &str) -> Option<String>;
 }
 
-impl BlockerSequenceExt for BlockerSequence {
+impl BlockersExt for Blockers {
 	fn current(&self) -> Option<&BlockerItem> {
 		last_item_in_list(&self.items)
 	}
@@ -202,13 +202,13 @@ mod tests {
 
 	#[test]
 	fn test_current() {
-		let seq = BlockerSequence::parse("- task 1\n- task 2\n- task 3");
+		let seq = Blockers::parse("- task 1\n- task 2\n- task 3");
 		assert_eq!(seq.current().map(|i| i.text.as_str()), Some("task 3"));
 	}
 
 	#[test]
 	fn test_current_skips_comments() {
-		let seq = BlockerSequence::parse("- task 1\n  comment\n- task 2\n  another comment");
+		let seq = Blockers::parse("- task 1\n  comment\n- task 2\n  another comment");
 		assert_eq!(seq.current().map(|i| i.text.as_str()), Some("task 2"));
 		// Comments should be attached to the item
 		assert_eq!(seq.current().map(|i| i.comments.len()), Some(1));
@@ -216,20 +216,20 @@ mod tests {
 
 	#[test]
 	fn test_current_with_context_no_hierarchy() {
-		let seq = BlockerSequence::parse("- Phase 1\n  - task 1\n- Phase 2\n  - task 2");
+		let seq = Blockers::parse("- Phase 1\n  - task 1\n- Phase 2\n  - task 2");
 		assert_eq!(seq.current_with_context(&[]), Some("Phase 2: task 2".to_string()));
 	}
 
 	#[test]
 	fn test_current_with_context_with_hierarchy() {
-		let seq = BlockerSequence::parse("- Phase 1\n  - task 1");
+		let seq = Blockers::parse("- Phase 1\n  - task 1");
 		let hierarchy = vec!["project".to_string()];
 		assert_eq!(seq.current_with_context(&hierarchy), Some("project: Phase 1: task 1".to_string()));
 	}
 
 	#[test]
 	fn test_current_with_context_multi_level_hierarchy() {
-		let seq = BlockerSequence::parse("- Section\n  - task");
+		let seq = Blockers::parse("- Section\n  - task");
 		let hierarchy = vec!["workspace".to_string(), "project".to_string()];
 		assert_eq!(seq.current_with_context(&hierarchy), Some("workspace: project: Section: task".to_string()));
 	}
@@ -237,7 +237,7 @@ mod tests {
 	#[test]
 	fn test_nested_items() {
 		let content = "- H1\n  - H2\n    - task under H2\n- Another H1\n  - task under another H1";
-		let seq = BlockerSequence::parse(content);
+		let seq = Blockers::parse(content);
 
 		// Current should be "task under another H1" with path "Another H1"
 		assert_eq!(seq.current_with_context(&[]), Some("Another H1: task under another H1".to_string()));
@@ -246,21 +246,21 @@ mod tests {
 	#[test]
 	fn test_deeply_nested() {
 		let content = "- Level 1\n  - Level 2\n    - Level 3\n      - deep task";
-		let seq = BlockerSequence::parse(content);
+		let seq = Blockers::parse(content);
 
 		assert_eq!(seq.current_with_context(&[]), Some("Level 1: Level 2: Level 3: deep task".to_string()));
 	}
 
 	#[test]
 	fn test_add() {
-		let mut seq = BlockerSequence::parse("- task 1");
+		let mut seq = Blockers::parse("- task 1");
 		seq.add("task 2");
 		assert_eq!(String::from(&seq), "- task 1\n- task 2");
 	}
 
 	#[test]
 	fn test_add_to_section() {
-		let mut seq = BlockerSequence::parse("- Section\n  - task 1");
+		let mut seq = Blockers::parse("- Section\n  - task 1");
 		seq.add("task 2");
 		// Should add under the same section (as sibling of task 1)
 		assert_eq!(String::from(&seq), "- Section\n  - task 1\n  - task 2");
@@ -268,7 +268,7 @@ mod tests {
 
 	#[test]
 	fn test_pop() {
-		let mut seq = BlockerSequence::parse("- task 1\n- task 2");
+		let mut seq = Blockers::parse("- task 1\n- task 2");
 		let popped = seq.pop(0);
 		assert_eq!(popped, Some("task 2".to_string()));
 		assert_eq!(String::from(&seq), "- task 1");
@@ -276,7 +276,7 @@ mod tests {
 
 	#[test]
 	fn test_pop_from_section() {
-		let mut seq = BlockerSequence::parse("- Section\n  - task 1\n  - task 2");
+		let mut seq = Blockers::parse("- Section\n  - task 1\n  - task 2");
 		let popped = seq.pop(0);
 		assert_eq!(popped, Some("task 2".to_string()));
 		assert_eq!(String::from(&seq), "- Section\n  - task 1");
@@ -284,7 +284,7 @@ mod tests {
 
 	#[test]
 	fn test_pop_empty() {
-		let mut seq = BlockerSequence::default();
+		let mut seq = Blockers::default();
 		let popped = seq.pop(0);
 		assert!(popped.is_none());
 	}
@@ -292,7 +292,7 @@ mod tests {
 	#[test]
 	fn test_serialize_roundtrip() {
 		let input = "- Header 1\n  - task 1\n- Header 2\n  - task 2";
-		let seq = BlockerSequence::parse(input);
+		let seq = Blockers::parse(input);
 		insta::assert_snapshot!(String::from(&seq), @"
 		- Header 1
 		  - task 1
@@ -303,17 +303,17 @@ mod tests {
 
 	#[test]
 	fn test_is_empty() {
-		let empty = BlockerSequence::default();
+		let empty = Blockers::default();
 		assert!(empty.is_empty());
 
-		let with_content = BlockerSequence::parse("- task");
+		let with_content = Blockers::parse("- task");
 		assert!(!with_content.is_empty());
 	}
 
 	#[test]
 	fn test_items_before_children() {
 		let content = "- root task\n- Section\n  - section task";
-		let seq = BlockerSequence::parse(content);
+		let seq = Blockers::parse(content);
 
 		// Current should be the section task
 		assert_eq!(seq.current_with_context(&[]), Some("Section: section task".to_string()));
@@ -322,7 +322,7 @@ mod tests {
 	#[test]
 	fn test_multiple_top_sections() {
 		let content = "- A\n  - task a\n- B\n  - task b\n- C\n  - task c";
-		let seq = BlockerSequence::parse(content);
+		let seq = Blockers::parse(content);
 
 		assert_eq!(seq.current_with_context(&[]), Some("C: task c".to_string()));
 
@@ -340,7 +340,7 @@ mod tests {
 	#[test]
 	fn test_comments_preserved() {
 		let content = "- task 1\n  comment 1\n  comment 2\n- task 2";
-		let seq = BlockerSequence::parse(content);
+		let seq = Blockers::parse(content);
 
 		assert_eq!(String::from(&seq), content);
 	}
@@ -348,13 +348,13 @@ mod tests {
 	#[test]
 	fn test_parse_and_serialize() {
 		let content = "- task 1\n  comment\n  - nested\n- task 2";
-		let seq = BlockerSequence::parse(content);
+		let seq = Blockers::parse(content);
 		assert_eq!(String::from(&seq), content);
 	}
 
 	#[test]
 	fn test_add_child_flat() {
-		let mut seq = BlockerSequence::parse("- task 1");
+		let mut seq = Blockers::parse("- task 1");
 		seq.add_child("subtask");
 		insta::assert_snapshot!(String::from(&seq), @"
 		- task 1
@@ -364,7 +364,7 @@ mod tests {
 
 	#[test]
 	fn test_add_child_to_section() {
-		let mut seq = BlockerSequence::parse("- Section\n  - task 1");
+		let mut seq = Blockers::parse("- Section\n  - task 1");
 		seq.add_child("subtask of task 1");
 		insta::assert_snapshot!(String::from(&seq), @"
 		- Section
@@ -375,14 +375,14 @@ mod tests {
 
 	#[test]
 	fn test_add_child_empty() {
-		let mut seq = BlockerSequence::default();
+		let mut seq = Blockers::default();
 		seq.add_child("first");
 		insta::assert_snapshot!(String::from(&seq), @"- first");
 	}
 
 	#[test]
 	fn test_set_flat() {
-		let mut seq = BlockerSequence::parse("- task 1\n- task 2");
+		let mut seq = Blockers::parse("- task 1\n- task 2");
 		let old = seq.set("replaced");
 		assert_eq!(old, Some("task 2".to_string()));
 		assert_eq!(String::from(&seq), "- task 1\n- replaced");
@@ -392,7 +392,7 @@ mod tests {
 	fn test_set_only_child_preserves_nesting() {
 		// The key case: current is the only child of its parent.
 		// pop+add would un-nest the replacement; set must keep it nested.
-		let mut seq = BlockerSequence::parse("- Section\n  - lonely task");
+		let mut seq = Blockers::parse("- Section\n  - lonely task");
 		seq.set("replacement");
 		insta::assert_snapshot!(String::from(&seq), @"
 		- Section
@@ -402,7 +402,7 @@ mod tests {
 
 	#[test]
 	fn test_set_deeply_nested() {
-		let mut seq = BlockerSequence::parse("- L1\n  - L2\n    - L3");
+		let mut seq = Blockers::parse("- L1\n  - L2\n    - L3");
 		seq.set("L3-new");
 		insta::assert_snapshot!(String::from(&seq), @"
 		- L1
@@ -413,21 +413,21 @@ mod tests {
 
 	#[test]
 	fn test_set_clears_comments() {
-		let mut seq = BlockerSequence::parse("- task\n  comment 1\n  comment 2");
+		let mut seq = Blockers::parse("- task\n  comment 1\n  comment 2");
 		seq.set("replaced");
 		assert_eq!(String::from(&seq), "- replaced");
 	}
 
 	#[test]
 	fn test_set_empty() {
-		let mut seq = BlockerSequence::default();
+		let mut seq = Blockers::default();
 		assert_eq!(seq.set("anything"), None);
 	}
 
 	#[test]
 	fn test_pop_with_one_parent_linear_chain() {
 		// farm tasks: shave yak: get a stool: remember where I left it
-		let mut seq = BlockerSequence::parse("- farm tasks\n  - shave yak\n    - get a stool\n      - remember where I left it");
+		let mut seq = Blockers::parse("- farm tasks\n  - shave yak\n    - get a stool\n      - remember where I left it");
 		let popped = seq.pop(1);
 		// Topmost popped item is "get a stool" (parent of leaf)
 		assert_eq!(popped, Some("get a stool".to_string()));
@@ -436,7 +436,7 @@ mod tests {
 
 	#[test]
 	fn test_pop_with_two_parents_linear_chain() {
-		let mut seq = BlockerSequence::parse("- farm tasks\n  - shave yak\n    - get a stool\n      - remember where I left it");
+		let mut seq = Blockers::parse("- farm tasks\n  - shave yak\n    - get a stool\n      - remember where I left it");
 		let popped = seq.pop(2);
 		assert_eq!(popped, Some("shave yak".to_string()));
 		assert_eq!(seq.current_with_context(&[]), Some("farm tasks".to_string()));
@@ -444,7 +444,7 @@ mod tests {
 
 	#[test]
 	fn test_pop_parents_exceeds_chain() {
-		let mut seq = BlockerSequence::parse("- only");
+		let mut seq = Blockers::parse("- only");
 		// chain depth = 0, asking for 1 parent → no-op, return None
 		let before: String = (&seq).into();
 		let popped = seq.pop(1);
@@ -455,13 +455,13 @@ mod tests {
 
 	#[test]
 	fn test_pop_parents_empty_sequence() {
-		let mut seq = BlockerSequence::default();
+		let mut seq = Blockers::default();
 		assert!(seq.pop(2).is_none());
 	}
 
 	#[test]
 	fn test_add_child_deeply_nested() {
-		let mut seq = BlockerSequence::parse("- L1\n  - L2\n    - L3");
+		let mut seq = Blockers::parse("- L1\n  - L2\n    - L3");
 		seq.add_child("L4");
 		insta::assert_snapshot!(String::from(&seq), @"
 		- L1
