+++ tedi/src/issue/types.rs
@@ -4,26 +4,10 @@
 
 use std::{collections::HashMap, path::PathBuf};
 
-use arrayvec::ArrayString;
-use copy_arrayvec::CopyArrayVec;
 use jiff::Timestamp;
 use serde::{Deserialize, Serialize};
+use tedi_core::{Blockers, IssueChildren, IssueIndex, IssueLink, IssueMarker, IssueSelector, Marker, ParseContext, ParseError, RepoInfo, TitleInGitPathError};
 use tedi_md::{indent_into, preserve_paragraph_spacing, wrap_inline_in_paragraphs};
-use url::Url;
-use v_utils::macros::wrap_err;
-
-/// Maximum title length enforced by Github.
-pub const MAX_TITLE_LENGTH: usize = 256;
-
-/// Maximum index depth (lineage + the issue itself).
-pub const MAX_INDEX_DEPTH: usize = MAX_LINEAGE_DEPTH + 1;
-pub const MAX_LINEAGE_DEPTH: usize = 8;
-pub type IssueChildren<T> = HashMap<IssueSelector, T>;
-use super::{
-	IssueMarker, Marker,
-	blocker::BlockerSequence,
-	error::{ParseContext, ParseError},
-};
 
 /// Trait for lazily loading an issue from a source.
 ///
@@ -54,75 +38,6 @@ pub trait LazyIssue<S: Clone + std::fmt::Debug>: Sized {
 		Ok(issue)
 	}
 }
-pub use tedi_adapters::github::RepoInfo;
-
-/// A Github issue identifier. Wraps a URL and derives all properties on demand.
-/// Format: `https://github.com/{owner}/{repo}/issues/{number}`
-#[derive(Clone, Debug, derive_more::Deref, derive_more::DerefMut, Eq, Hash, PartialEq)]
-pub struct IssueLink(Url);
-
-impl IssueLink /*{{{1*/ {
-	/// Create from a URL. Returns None if not a valid Github issue URL.
-	pub fn new(mut url: Url) -> Option<Self> {
-		// Validate it's a Github issue URL
-		if url.host_str() != Some("github.com") {
-			return None;
-		}
-		let segments: Vec<_> = url.path_segments()?.collect();
-		// Must be: owner/repo/issues/number
-		if segments.len() < 4 || segments[2] != "issues" {
-			return None;
-		}
-		// Number must be valid
-		segments[3].parse::<u64>().ok()?;
-		// Normalize owner/repo to lowercase: Github treats them case-insensitively,
-		// so every reader (owner(), repo(), as_str()) sees the same canonical form.
-		let normalized: Vec<String> = segments.iter().enumerate().map(|(i, s)| if i < 2 { s.to_lowercase() } else { s.to_string() }).collect();
-		url.path_segments_mut().ok()?.clear().extend(normalized.iter().map(String::as_str));
-		Some(Self(url))
-	}
-
-	/// Parse from a URL string.
-	pub fn parse(url: &str) -> Option<Self> {
-		let url = Url::parse(url).ok()?;
-		Self::new(url)
-	}
-
-	/// Get the underlying URL.
-	pub fn url(&self) -> &Url {
-		&self.0
-	}
-
-	/// Get the repository info (owner and repo).
-	pub fn repo_info(&self) -> RepoInfo {
-		let mut segments = self.0.path_segments().unwrap();
-		let owner = segments.next().unwrap();
-		let repo = segments.next().unwrap();
-		RepoInfo::new(owner, repo)
-	}
-
-	/// Get the owner (first path segment).
-	pub fn owner(&self) -> &str {
-		self.0.path_segments().unwrap().next().unwrap()
-	}
-
-	/// Get the repo (second path segment).
-	pub fn repo(&self) -> &str {
-		self.0.path_segments().unwrap().nth(1).unwrap()
-	}
-
-	/// Get the issue number (fourth path segment).
-	pub fn number(&self) -> u64 {
-		self.0.path_segments().unwrap().nth(3).unwrap().parse().unwrap()
-	}
-
-	/// Build URL string.
-	pub fn as_str(&self) -> &str {
-		self.0.as_str()
-	}
-}
-//,}}}1
-
 /// Identity of a comment - either linked to Github or pending creation.
 /// Note: The first comment (issue body) is always `Body`, not `Linked` or `Pending`.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
@@ -468,7 +383,7 @@ impl IssueIdentity {
 	///
 	/// # Errors
 	/// Returns `TitleInGitPathError` if any parent selector is a Title (pending issue).
-	pub fn git_lineage(&self) -> Result<Vec<u64>, super::error::TitleInGitPathError> {
+	pub fn git_lineage(&self) -> Result<Vec<u64>, TitleInGitPathError> {
 		self.parent_index.git_num_path()
 	}
 
@@ -479,189 +394,19 @@ impl IssueIdentity {
 	}
 }
 
-/// Selector for identifying an issue within a repo.
-/// GitId is preferred when available, as title can change.
-/// Uses `ArrayString` for `Copy` semantics.
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-#[allow(clippy::large_enum_variant)] // Intentional: Title variant is large for Copy semantics
-pub enum IssueSelector {
-	/// Github issue number (stable identifier)
-	GitId(u64),
-	/// Issue title (for pending issues not yet synced to Github)
-	Title(ArrayString<MAX_TITLE_LENGTH>),
-	/// Regex pattern for fuzzy matching (touch mode only, lower priority than Title)
-	Regex(ArrayString<MAX_TITLE_LENGTH>),
-}
-
-impl IssueSelector {
-	/// Create a Title selector from a string.
-	/// Panics if title exceeds MAX_TITLE_LENGTH (256 chars).
-	pub fn title(title: &str) -> Self {
-		Self::Title(ArrayString::from(title).unwrap_or_else(|_| panic!("title too long (max {MAX_TITLE_LENGTH} chars): {}", title.len())))
-	}
-
-	/// Try to create a Title selector from a string.
-	/// Returns None if title exceeds MAX_TITLE_LENGTH.
-	pub fn try_title(title: &str) -> Option<Self> {
-		ArrayString::from(title).ok().map(Self::Title)
-	}
-
-	/// Create a Regex selector from a pattern string.
-	pub fn regex(pattern: &str) -> Self {
-		Self::Regex(ArrayString::from(pattern).unwrap_or_else(|_| panic!("pattern too long (max {MAX_TITLE_LENGTH} chars): {}", pattern.len())))
-	}
-}
-
-/// Minimal descriptor for locating an issue.
-/// Contains repo info and a path of selectors from root to the target issue (inclusive).
-/// Uses fixed-size storage to be `Copy`.
-#[derive(Clone, Copy, Debug, derive_more::Deref, derive_more::DerefMut, Eq, PartialEq)]
-pub struct IssueIndex {
-	repo_info: RepoInfo,
-	/// Path from root to target issue (inclusive).
-	#[deref]
-	#[deref_mut]
-	index: CopyArrayVec<IssueSelector, MAX_INDEX_DEPTH>,
-}
-
-impl IssueIndex {
-	/// Create descriptor for a root-level issue.
-	pub fn root(repo_info: RepoInfo, selector: IssueSelector) -> Self {
-		let mut index = CopyArrayVec::new();
-		index.push(selector);
-		Self { repo_info, index }
-	}
-
-	/// Create descriptor with full index path.
-	/// Panics if index exceeds MAX_INDEX_DEPTH.
-	pub fn with_index(repo_info: RepoInfo, index: Vec<IssueSelector>) -> Self {
-		Self {
-			repo_info,
-			index: index.into_iter().collect(),
-		}
-	}
-
-	/// Create descriptor for repo only (no specific issue).
-	pub fn repo_only(repo_info: RepoInfo) -> Self {
-		Self {
-			repo_info,
-			index: CopyArrayVec::new(),
-		}
-	}
-
-	/// Add a child selector, returning new descriptor.
-	/// Panics if result would exceed MAX_INDEX_DEPTH.
-	pub fn child(&self, selector: IssueSelector) -> Self {
-		let mut index = self.index;
-		index.push(selector);
-		Self { repo_info: self.repo_info, index }
-	}
-
-	/// Get the index path.
-	pub fn index(&self) -> &[IssueSelector] {
-		&self.index
-	}
-
-	/// Get the repository info.
-	pub fn repo_info(&self) -> RepoInfo {
-		self.repo_info
-	}
-
-	/// Get the owner.
-	pub fn owner(&self) -> &str {
-		self.repo_info.owner()
-	}
-
-	/// Get the repo.
-	pub fn repo(&self) -> &str {
-		self.repo_info.repo()
-	}
-
-	/// Extract numeric issue numbers from the index (GitId selectors only).
-	///
-	/// # Errors
-	/// Returns `TitleInGitPathError` if any selector is a Title (pending issue).
-	pub fn git_num_path(&self) -> Result<Vec<u64>, super::error::TitleInGitPathError> {
-		use miette::{NamedSource, SourceSpan};
-
-		let mut result = Vec::with_capacity(self.index().len());
-		let mut offset = format!("{}/{}", self.repo_info.owner(), self.repo_info.repo()).len();
-
-		for selector in self.index() {
-			match selector {
-				IssueSelector::GitId(n) => {
-					let s = format!("/{n}");
-					offset += s.len();
-					result.push(*n);
-				}
-				IssueSelector::Title(title) | IssueSelector::Regex(title) => {
-					let span: SourceSpan = (offset + 1, title.len()).into(); // +1 to skip the '/'
-					return Err(super::error::TitleInGitPathError::new(NamedSource::new("IssueIndex", self.to_string()), span));
-				}
+/// Marker view of an issue's identity — how it serializes to a title-line marker.
+impl From<&IssueIdentity> for IssueMarker {
+	fn from(identity: &IssueIdentity) -> Self {
+		if let Some(meta) = identity.as_linked() {
+			IssueMarker::Linked {
+				user: meta.user.clone(),
+				link: meta.link().clone(),
 			}
-		}
-		Ok(result)
-	}
-
-	/// Get the issue's own number if the last selector is a GitId.
-	pub fn issue_number(&self) -> Option<u64> {
-		match self.index().last() {
-			Some(IssueSelector::GitId(n)) => Some(*n),
-			_ => None,
-		}
-	}
-
-	/// Get the parent's IssueIndex (all selectors except the last one).
-	/// For repo-only or single-selector indices, returns repo_only.
-	pub fn parent(&self) -> Option<Self> {
-		if self.index.is_empty() {
-			None
+		} else if identity.is_virtual {
+			IssueMarker::Virtual
 		} else {
-			let mut index = self.index;
-			index.pop();
-			Some(Self { repo_info: self.repo_info, index })
+			IssueMarker::Pending
 		}
-	}
-}
-
-impl std::fmt::Display for IssueIndex {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		write!(f, "{}/{}", self.repo_info.owner(), self.repo_info.repo())?;
-		for selector in self.index() {
-			match selector {
-				IssueSelector::GitId(n) => write!(f, "/{n}")?,
-				IssueSelector::Title(t) | IssueSelector::Regex(t) => write!(f, "/{t}")?,
-			}
-		}
-		Ok(())
-	}
-}
-
-/// Error returned when parsing an `IssueIndex` from a string fails.
-#[wrap_err]
-#[derive(Debug, thiserror::Error)]
-#[error("{msg}")]
-pub struct IssueIndexParseError {
-	msg: String,
-}
-
-impl std::str::FromStr for IssueIndex {
-	type Err = IssueIndexParseError;
-
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		let parts: Vec<&str> = s.split('/').collect();
-		if parts.len() < 2 {
-			return Err(IssueIndexParseError::new(format!("IssueIndex requires at least owner/repo, got: {s}")));
-		}
-		let repo_info = RepoInfo::new(parts[0], parts[1]);
-		let selectors: Vec<IssueSelector> = parts[2..]
-			.iter()
-			.map(|p| match p.parse::<u64>() {
-				Ok(n) => IssueSelector::GitId(n),
-				Err(_) => IssueSelector::title(p),
-			})
-			.collect();
-		Ok(Self::with_index(repo_info, selectors))
 	}
 }
 
@@ -762,7 +507,7 @@ pub struct IssueContents {
 	pub labels: Vec<String>,
 	pub state: CloseState,
 	pub comments: Comments,
-	pub blockers: BlockerSequence,
+	pub blockers: Blockers,
 }
 /// Complete representation of an issue file
 #[derive(Clone, Debug, PartialEq)]
@@ -914,7 +659,7 @@ impl Issue /*{{{1*/ {
 	///
 	/// # Errors
 	/// Returns `TitleInGitPathError` if any parent selector is a Title (pending issue).
-	pub fn lineage(&self) -> Result<Vec<u64>, super::error::TitleInGitPathError> {
+	pub fn lineage(&self) -> Result<Vec<u64>, TitleInGitPathError> {
 		self.identity.git_lineage()
 	}
 
@@ -951,7 +696,7 @@ impl Issue /*{{{1*/ {
 	/// # Errors
 	/// Returns `IssueError::ErroneousComposition` if a git-linked child (GitId selector) exists
 	/// in virtual but not in hollow — either an internal bug or user manually embedded a linked marker.
-	pub fn from_combined(hollow: HollowIssue, virtual_issue: VirtualIssue, parent_idx: IssueIndex, is_virtual: bool) -> Result<Self, super::error::IssueError> {
+	pub fn from_combined(hollow: HollowIssue, virtual_issue: VirtualIssue, parent_idx: IssueIndex, is_virtual: bool) -> Result<Self, tedi_core::IssueError> {
 		let identity = IssueIdentity {
 			parent_index: parent_idx,
 			is_virtual,
@@ -965,7 +710,7 @@ impl Issue /*{{{1*/ {
 				Some(ch) => ch.clone(),
 				None => {
 					if let IssueSelector::GitId(n) = selector {
-						return Err(super::error::IssueError::new_erroneous_composition(
+						return Err(tedi_core::IssueError::new_erroneous_composition(
 							n,
 							"either internal bug (HollowIssue was constructed incorrectly) or user manually embedded a `<!-- @user url -->` marker, which is not permitted".to_string(),
 						));
@@ -1665,7 +1410,7 @@ impl VirtualIssue {
 								"non-checkbox list after blockers section".into(),
 							));
 						}
-						// Blocker list — collect events for BlockerSequence parsing
+						// Blocker list — collect events for Blockers parsing
 						let list_end = Self::find_matching_end_list(events, pos);
 						blocker_events.extend(events[pos..list_end].iter().cloned());
 						blocker_list_consumed = true;
@@ -1732,11 +1477,11 @@ impl VirtualIssue {
 
 		// Parse blockers from collected events
 		let blockers = if blocker_events.is_empty() {
-			BlockerSequence::default()
+			Blockers::default()
 		} else {
-			// Render blocker events to string and parse with BlockerSequence::parse
+			// Render blocker events to string and parse with Blockers::parse
 			let blocker_text: String = super::Events::from(blocker_events).into();
-			BlockerSequence::parse(&blocker_text)
+			Blockers::parse(&blocker_text)
 		};
 
 		Ok(VirtualIssue {
@@ -1749,7 +1494,7 @@ impl VirtualIssue {
 				blockers: {
 					let mut seq = blockers;
 					if select_blockers {
-						seq.set_state = Some(crate::issue::BlockerSetState::Pending);
+						seq.set_state = Some(tedi_core::BlockerSetState::Pending);
 					}
 					seq
 				},
@@ -1920,31 +1665,6 @@ mod tests {
 		};
 		let children = v.children.iter().map(|(sel, child)| (*sel, hollow_from_virtual(child))).collect();
 		HollowIssue::new(remote, children)
-	}
-
-	#[test]
-	fn repo_info_lowercases_owner_and_repo() {
-		let info = RepoInfo::new("MyOrg", "MyRepo");
-		assert_eq!(info.owner(), "myorg");
-		assert_eq!(info.repo(), "myrepo");
-
-		// The `From` and `FromStr` paths route through `new`, so they normalize too.
-		let from_tuple: RepoInfo = ("FooBar", "Baz-Qux").into();
-		assert_eq!((from_tuple.owner(), from_tuple.repo()), ("foobar", "baz-qux"));
-		let idx: IssueIndex = "OWNER/REPO/123".parse().unwrap();
-		assert_eq!((idx.repo_info().owner(), idx.repo_info().repo()), ("owner", "repo"));
-	}
-
-	#[test]
-	fn issue_link_normalizes_owner_and_repo() {
-		let link = IssueLink::parse("https://github.com/MyOrg/MyRepo/issues/42").unwrap();
-		// Every reader sees the canonical lowercase form, including the raw URL string.
-		assert_eq!(link.owner(), "myorg");
-		assert_eq!(link.repo(), "myrepo");
-		assert_eq!(link.as_str(), "https://github.com/myorg/myrepo/issues/42");
-		assert_eq!((link.repo_info().owner(), link.repo_info().repo()), ("myorg", "myrepo"));
-		// The issue number segment is untouched (no case to normalize, but ensure it's preserved).
-		assert_eq!(link.number(), 42);
 	}
 
 	#[test]
@@ -2206,7 +1926,7 @@ mod tests {
 	fn test_select_blockers_standalone_line() {
 		let content = "- [ ] Issue <!-- https://github.com/owner/repo/issues/1 -->\n\n  Body\n\n  !s\n  # Blockers\n  - task 1\n  - task 2\n";
 		let vi = VirtualIssue::parse(content, PathBuf::from("test.md")).unwrap();
-		assert!(matches!(vi.contents.blockers.set_state, Some(crate::issue::BlockerSetState::Pending)));
+		assert!(matches!(vi.contents.blockers.set_state, Some(tedi_core::BlockerSetState::Pending)));
 		// !s must not appear in re-serialized output, blockers preserved
 		insta::assert_snapshot!(unsafe_mock_parse_virtual(content).serialize_virtual(), @"
 		- [ ] Issue <!-- https://github.com/owner/repo/issues/1 -->
@@ -2222,7 +1942,7 @@ mod tests {
 	fn test_select_blockers_suffix_on_header() {
 		let content = "- [ ] Issue <!-- https://github.com/owner/repo/issues/1 -->\n\n  Body\n\n  # Blockers !s\n  - one\n";
 		let vi = VirtualIssue::parse(content, PathBuf::from("test.md")).unwrap();
-		assert!(matches!(vi.contents.blockers.set_state, Some(crate::issue::BlockerSetState::Pending)));
+		assert!(matches!(vi.contents.blockers.set_state, Some(tedi_core::BlockerSetState::Pending)));
 		// blocker text parsed correctly despite !s suffix on header
 		insta::assert_snapshot!(vi.contents.blockers.items[0].text, @"one");
 	}
@@ -2240,7 +1960,7 @@ mod tests {
 		let content = "- [ ] Issue <!-- https://github.com/owner/repo/issues/1 -->\n\n  !S\n  # Blockers\n  - task\n";
 		let virtual_issue = VirtualIssue::parse(content, PathBuf::from("test.md")).unwrap();
 
-		assert!(matches!(virtual_issue.contents.blockers.set_state, Some(crate::issue::BlockerSetState::Pending)));
+		assert!(matches!(virtual_issue.contents.blockers.set_state, Some(tedi_core::BlockerSetState::Pending)));
 	}
 
 	#[test]
