+++ tedi_core/src/error.rs
